introGame/newPikiGame progress

+++ include/FlowController.h
@@ -33,7 +33,7 @@ struct FlowController {
 	char mStagePath2[0x80];        // _1B0, identical to above
 	int mNaviOnMap;                // _230, unknown
 	int _234;                      // _234
-	u8 _238;                       // _238
+	u32 _238;                      // _238
 	u32 _23C;                      // _23C
 	int _240;                      // _240
 	int _244;                      // _244
